requirements.txt
sparkfun-qwiic-sgp40
sparkfun-qwiic-i2c
